+++ lib/document_detector_sdk.dart
@@ -1,11 +1,18 @@
 library document_detector_sdk;
 
+import 'dart:async';
+import 'dart:convert';
+import 'dart:io';
+
+import 'package:dio/dio.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/painting.dart';
 import 'package:flutter/services.dart';
 
 part 'src/document_detector.dart';
+part 'src/document_detector_layout.dart';
 part 'src/document_detector_result.dart';
 part 'src/document_detector_sdk.dart';
 part 'src/document_type.dart';
+part 'src/image_response.dart';
 part 'src/sdk_result.dart';