+++ lib/src/document_detector_layout.dart
@@ -1,0 +1,43 @@
+part of document_detector_sdk;
+
+class DocumentDetectorLayout {
+  final String closeImageName;
+  final String greenMaskImageName;
+  final String whiteMaskImageName;
+  final String redMaskImageName;
+  final String soundOnImageName;
+  final String soundOffImageName;
+
+  DocumentDetectorLayout(
+      {this.closeImageName,
+      this.greenMaskImageName,
+      this.whiteMaskImageName,
+      this.redMaskImageName,
+      this.soundOnImageName,
+      this.soundOffImageName});
+
+  Map<String, dynamic> toMap() {
+    Map<String, dynamic> ret = Map<String, dynamic>();
+
+    if (greenMaskImageName != null) {
+      ret['greenMaskImageName'] = greenMaskImageName;
+    }
+    if (whiteMaskImageName != null) {
+      ret['whiteMaskImageName'] = whiteMaskImageName;
+    }
+    if (redMaskImageName != null) {
+      ret['redMaskImageName'] = redMaskImageName;
+    }
+    if (soundOnImageName != null) {
+      ret['soundOnImageName'] = soundOnImageName;
+    }
+    if (soundOffImageName != null) {
+      ret['soundOffImageName'] = soundOffImageName;
+    }
+    if (closeImageName != null) {
+      ret['closeImageName'] = closeImageName;
+    }
+
+    return ret;
+  }
+}