+++ lib/src/image_response.dart
@@ -1,0 +1,13 @@
+part of document_detector_sdk;
+
+class ImageResponse {
+  final String uploadUrl;
+  final String getUrl;
+
+  ImageResponse({@required this.uploadUrl, @required this.getUrl});
+
+  @override
+  String toString() {
+    return '$uploadUrl, $getUrl';
+  }
+}